fix: cleanup accounts helper functions
