+++ api/ContextAccounts.tsx
@@ -4,11 +4,13 @@ import {
   transformNotesVersion,
 } from "@/components/ui-elements/notes-writer/NotesWriter";
 import { toast } from "@/components/ui/use-toast";
+import {
+  calcOrder,
+  getQuotaFromTerritoryOrSubsidaries,
+} from "@/helpers/accounts";
 import { calcPipeline } from "@/helpers/projects";
 import { SelectionSet, generateClient } from "aws-amplify/data";
-import { differenceInDays } from "date-fns";
-import { max } from "lodash";
-import { find, flow, get, map, sortBy, sum } from "lodash/fp";
+import { flow, map, sortBy } from "lodash/fp";
 import { FC, ReactNode, createContext, useContext } from "react";
 import useSWR from "swr";
 import { handleApiErrors } from "./globals";
@@ -94,39 +96,10 @@ const selectionSet = [
   "payerAccounts.awsAccountNumber",
 ] as const;
 
-type AccountData = SelectionSet<Schema["Account"]["type"], typeof selectionSet>;
-type TerritoryData = AccountData["territories"][number];
-type SubsidiaryData = AccountData["subsidiaries"][number];
-type ResponsibilityData =
-  TerritoryData["territory"]["responsibilities"][number];
-
-const getLatestQuota = (territories: TerritoryData[]) =>
-  flow(
-    map(
-      (t: TerritoryData) =>
-        flow(
-          sortBy((r: ResponsibilityData) => -new Date(r.startDate).getTime()),
-          find((r) => differenceInDays(new Date(), new Date(r.startDate)) > 0),
-          get("quota")
-        )(t.territory.responsibilities) || 0
-    ),
-    sum
-  )(territories);
-
-const getQuotaFromTerritoryOrSubsidaries = (
-  territories: TerritoryData[],
-  subsidiaries: SubsidiaryData[]
-) =>
-  max([
-    getLatestQuota(territories),
-    flow(
-      map((s: SubsidiaryData) => getLatestQuota(s.territories)),
-      sum
-    )(subsidiaries),
-  ]) || 0;
-
-export const calcOrder = (quota: number, pipeline: number): number =>
-  sum([Math.floor(quota / 1000) * 1000, Math.floor(pipeline / 1000)]);
+export type AccountData = SelectionSet<
+  Schema["Account"]["type"],
+  typeof selectionSet
+>;
 
 const mapAccount: (account: AccountData) => Account = ({
   id: accountId,
